+++ lib/ferry/data/models/station/station.model.dart
@@ -15,4 +15,26 @@ class Station with _$Station {
     required int id,
     required String name,
   }) = _Station;
+
+  String? get backgroundAssetPath {
+    switch (id) {
+      case 1:
+        return 'assets/locations/konak.jpg';
+      case 2:
+        return 'assets/locations/karsiyaka.jpg';
+      case 3:
+        return 'assets/locations/bostanli.jpg';
+      case 4:
+        return 'assets/locations/pasaport.jpg';
+      case 5:
+        return 'assets/locations/alsancak.jpg';
+      case 6:
+        return 'assets/locations/goztepe.jpg';
+      case 7:
+        return 'assets/locations/uckuyular.jpg';
+      case 1019:
+        return 'assets/locations/karantina.jpg';
+    }
+    return null;
+  }
 }