+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### For Users
+
+### For Developers
+
+#### Changed
+
+ - Add `Station.backgroundAssetPath` and refactor references
+
 ## [0.2.0] - 24.05.2023
 
 ### Added